thread: move the yield flag and the preemption level to the run queue
There is really no need to make the yield flag and the preemption level
thread-local on a single processor system, so turn them into globals,
as members of the unique run queue. This slightly improves preemption
control functions by removing the indirection of the current thread.

Another motivation for this change is that X15 is already an example
of a thread-local preemption counter, because of SMP. As a result,
it made sense to show another way, which is actually the common one
on small single processor embedded systems.